✨Improve: コントローラの共通メソッドをconcernsに移譲
